+++ CHANGES.md
@@ -11,6 +11,12 @@
   - `print`, `print-root`
   - `->string`, `->root-string`
   - `append-newline`, `prepend-newline`
+  - `right*`, `left*`, ... (delegating to `fast-zip.core/right`, ...)
+- new token type `:forms`
+- new functions in `rewrite-clj.parser`:
+  - `parse-all`
+  - `parse-string-all`
+  - `parse-file-all`
 
 ### 0.1.0